cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hopx-cli"
version = "0.1.0"
description = "Official CLI for Hopx.ai cloud sandboxes"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "HOPX.AI", email = "support@hopx.ai"}
]
keywords = ["hopx", "hopx-ai", "cli", "sandbox", "vm", "microvm", "cloud"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Shells",
    "Typing :: Typed",
]

dependencies = [
    "hopx-ai>=0.3.8",
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "keyring>=24.0.0",
    "pyyaml>=6.0.0",
    "click>=8.0.0",
    "pyperclip>=1.8.0",
    "qrcode>=7.4.0",
    "packaging>=23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.3.0",
    "ruff>=0.8.0",
    "mypy>=1.8.0",
    "types-pyyaml>=6.0.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
hopx = "hopx_cli.main:app"

[project.urls]
Homepage = "https://hopx.ai"
Documentation = "https://docs.hopx.ai/cli"
Repository = "https://github.com/hopx-ai/hopx"
Issues = "https://github.com/hopx-ai/hopx/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/hopx_cli"]

[tool.hatch.build.targets.sdist]
include = [
    "src/**/*.py",
    "src/hopx_cli/py.typed",
]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "W",      # pycodestyle warnings
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "ARG",    # flake8-unused-arguments
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "E402",   # module level import not at top of file (typer pattern)
    "B008",   # function call in argument defaults (typer.Option pattern)
    "B904",   # raise from (not useful for typer.Exit)
    "SIM105", # contextlib.suppress (try/except pass is clearer)
    "SIM117", # nested with statements (clearer in some cases)
    "UP007",  # use X | Y for Union types
    "UP045",  # use X | None for Optional
    "UP047",  # use type parameters (TypeVar is still valid)
    "RUF022", # unsorted __all__ (intentional grouping)
]

[tool.ruff.lint.isort]
known-first-party = ["hopx_cli"]

[tool.ruff.lint.per-file-ignores]
"src/hopx_cli/commands/*.py" = ["ARG001"]  # typer commands often have unused ctx
"src/hopx_cli/main.py" = ["ARG001"]  # typer version callback pattern
"tests/**/*.py" = ["ARG001", "ARG002"]  # test fixtures often have unused args
"tests/*.py" = ["ARG001", "ARG002", "E722"]  # test files

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "-ra",
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "unit: Pure unit tests (no I/O, fast)",
    "integration: Integration tests (may use mocked I/O)",
    "slow: Tests taking >1 second",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:typer.*",
    "ignore::DeprecationWarning:click.*",
    "ignore::DeprecationWarning:websockets.*",
    "ignore::DeprecationWarning:hopx_ai.*",
]

[tool.coverage.run]
source = ["src/hopx_cli"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
    "def __repr__",
    "\\.\\.\\.",
]
show_missing = true
fail_under = 45
